trajectory_visualization/requirements.txt
matplotlib>=3.3.4
numpy>=1.13.3
PyYAML>=6.0
scipy>=1.5.4
